finished linking db w/ ui

+++ MainModule/src/Counselor/StudentScore.java
@@ -3,9 +3,9 @@ package Counselor;
 public class StudentScore {
     private String parentName;
     private String studentName;
-    private double studentScore;
+    private int studentScore;
 
-    StudentScore(String parent, String studentName, double studentScore) {
+    StudentScore(String parent, String studentName, int studentScore) {
         this.parentName = parent;
         this.studentName = studentName;
         this.studentScore = studentScore;
@@ -17,5 +17,5 @@ public class StudentScore {
     public String getStudentName() {
         return this.studentName;
     }
-    public double getStudentScore() { return this.studentScore; }
+    public int getStudentScore() { return this.studentScore; }
 }
